First commit
I've started on the layout of the map and have done pseudocode for Players, Game Manager, Tile Events, and Tile Info.

Since this is my third third time restarting, I'm trying to implement some things that will make coding a lot easier such as namespaces, abstract   and interface classes, and a much smoother way to transition between each step of the board game.