Players Prog
Made Move() and UseItem() as abstract functions so I can make override them in P0, P1, etc. They will be able to reference their own transform.position and I'll be able to move them.

Also added an int endTile to the abstract Players class, and a setEndTile(...) function to be used in P0, P1, etc.

In P0: Overrided the Move() function and it should be working. Need to implement the dice roll so it can change bool onEndTile to true when currTile == endTile, thus stopping the GM from running Move().

Going to start the dice roll class next.