Tile Revamp
Figured out how to generically run "RunTileEvent()" and also find the number of components in a gameobject. This will allow me to figure out if the tile has a component/special tile event if number of components > 1 (It will always have at least 1 because of the Transform component), then I can assign a var type variable to the component in the 1st spot of the component list which will be the special tile script.

**Need to fix line 53-54 in GameManager.cs \\ Put Data.playersArr[Data.currPlayer].getCurrTile() where the 0 is and replace it with 1 for now.

If # of components is == 1, then skip over this section since it's a normal blank tile (Which I probably won't have )