GameManager TileEvents and Dice Animation
Alright! I finished up the RunTileEvent() function in GameManager.cs, so whenever I call that, it will grab a list of the components of the current tile, check if it has more than just the transform component, check if the tile is a START/TROPHY or onEndTile == true, and if it satisfies all of those conditions it will run the event.

For each of the tile event's individual scripts, I added a GetName() function, so I grab that to ensure the type of tile is what I'm looking for later on.

Starting to implement the dice! As of right now, I animated the dice roll, so it will "roll" from 1,2,3,4,5,6 on a loop. I'll add variation later on.

> What I'm hoping to do with this is have the animation run, and when the player stops the dice, it stops the animation and I get the actual name of the sprite to pull the value that the player rolled. I'm not exactly sure if this will work, but that's what I have in mind.